[avs_commons] HTTP client
Summary:
I might make some more changes to this diff, but it's already in a working state. The changes I might do are mostly related to what parts of the code might or might not be duplicated between Commons and libCWMP.

The first revision is just moving files from libCWMP. The second revision contains all the actual changes, and there are quite a few. So you might review just the changes, or all the code from scratch. I give you the choice ;)

I have a working but somewhat rough migration of libCWMP to this. I'll polish and upload it tomorrow.

Test Plan: make check; also make check in libcwmp

Reviewers: marian, mkrawiec

Reviewed By: marian, mkrawiec

Differential Revision: https://phabricator.avsystem.com/D5124