Add support for `signed char` in `ASSERT_EQ`
Attempt to use `signed char` in `AVS_ASSERT_EQ` failed, because we took
care to handle all types except for `signed char`, foolishly believing
that `char` and `unsigned char` is enough.

Of course it isn't. It's C after all.

In C and C++, `char` may be either signed or unsigned, which does not
mean that `char` is ever fully equivalent to `signed char` or `unsigned
char`. No, each one is a completely separate type, unlike every single
other integral type, which is either implicitly signed, or explicitly
unsigned.

When attempting to force the compiler to display any kind of warning
whenever an implicit `enum -> int` conversion is performed, I enabled
`-fshort-enums` GCC flag, which causes all enums to be backed by
smallest possible numeric type able to hold all enum values instead of
`int`. Which made the `ASSERT_EQ` fail when it was used to compare
a two-value enum that may be equal to either -1 or 0. After enabling
this flag, the best numeric type GCC was able to use of course was not
`char`, oh no - we can't be sure whether it's signed or not after all.
The best one was `signed char`. Which caused an oh-so-helpful error
message:

`error: invalid use of void expression`

`signed char`, I hate you so much.